aider: Removed lint settings for `yamllint` from ale.vim configuration file.
# Aider chat conversation:

USER: If coc is handling the linting, then ale shouldn't have the lint settings. Please remove the lint settings from ale.vim

ASSISTANT: Here is the diff to remove the lint settings for `yamllint` from the `ale.vim` configuration file, as `coc.nvim` is handling the linting:

```diff
--- /home/harleypig/projects/dotvim/.vim/pack/settings/start/settings/plugin/ale.vim
+++ /home/harleypig/projects/dotvim/.vim/pack/settings/start/settings/plugin/ale.vim
@@ ... @@
 " Auto cleanup for all files
 let g:ale_fixers = {
 \   '*': ['remove_trailing_lines', 'trim_whitespace'],
 \}

-let g:ale_yaml_yamllint_executable = 'yamllint'
-let g:ale_yaml_yamllint_options = '-d "{extends: default, rules: {line-length: {max: 120}}}"'
-
-let g:ale_linters = {
-  \   'yaml': ['yamllint'],
-  \}

 " Map ALEFix
 map <leader>fb :ALEFix<CR>
```

This change removes the configuration lines that set up `yamllint` as a linter in ALE, as well as the linter settings for YAML files.