aider: Added configuration for `eslint` in the `.vim/coc-settings.json` file to enable linting for JavaScript, TypeScript, and other specified file types.
# Aider chat conversation:

USER: Does eslint need to be configured in coc?

ASSISTANT: Yes, `eslint` should be configured in CoC if you are using it for linting JavaScript and other ECMAScript-based languages like TypeScript. CoC (Conquer of Completion) is an intellisense engine for Vim that can provide linting, code completion, and other language features when properly configured with language servers or extensions.

To configure `eslint` with CoC, you would typically ensure that the `coc-eslint` extension is installed and then configure it within the `.vim/coc-settings.json` file in your Vim configuration directory. This file is used to set up various CoC extensions and their settings.

If you would like to proceed with configuring `eslint` in CoC, please add the `.vim/coc-settings.json` file to the chat, and I will provide you with the necessary changes.

USER: I added these *read-write* files: .vim/coc-settings.json

ASSISTANT: To configure `eslint` in the `.vim/coc-settings.json` file, we will add the necessary configuration for the `coc-eslint` extension. Here is the diff for the changes:

```diff
--- .vim/coc-settings.json
+++ .vim/coc-settings.json
@@ ... @@
   // Pyright settings
   // https://github.com/fannheyward/coc-pyright
   // https://github.com/microsoft/pyright
   // https://microsoft.github.io/pyright
   // look in pyrightconfig.json or pyproject.toml for more settings

   "pyright.inlayHints.functionReturnTypes": false,
   "pyright.inlayHints.variableTypes": false,
+
+  // eslint settings
+  // Requires coc-eslint extension to be installed
+  "eslint.enable": true,
+  "eslint.packageManager": "npm",
+  "eslint.autoFixOnSave": true,
+  "eslint.filetypes": ["javascript", "javascriptreact", "typescript", "typescriptreact", "vue", "html"],
+
   // coc-pyright flake8
   "python.linting.flake8Enabled": true,
   "python.linting.flake8Stdin": true,
```

This configuration enables `eslint`, sets it to use `npm` as the package manager, enables automatic fixing on save, and specifies the file types that `eslint` should lint. Make sure you have the `coc-eslint` extension installed for these settings to take effect.