aider: Update ESLint settings in `.vim/coc-settings.json` to specify file types for ESLint to apply to.
# Aider chat conversation:

USER: Why is vim reporting an error when trying to save a non-json file? Error detected while processing BufWrite Autocommands for "*"..function coc#rpc#request[4]..<SNR>35_request[34]..BufWrite Autocommands for "*"..function coc#rpc#request[4]..<SNR>35_r
equest:
line   29:
E605: Exception not caught: Error on "doAutocmd" request: Request textDocument/codeAction failed with message: No ESLint configuration found in /home/harleypig/projects/dotvim.
Press ENTER or type command to continue

ASSISTANT: The error message you're encountering suggests that ESLint is being triggered for a non-JSON file and is unable to find a configuration file for ESLint in the specified directory. This could be due to the ESLint settings in the CoC configuration which are applied globally to all file types.

To resolve this issue, we need to adjust the ESLint settings in the `.vim/coc-settings.json` file to ensure that ESLint is only triggered for file types that it should be responsible for, such as JavaScript or TypeScript files.

Here is the diff to update the `.vim/coc-settings.json` file to include a specific configuration for ESLint to only apply to certain file types:

```diff
--- .vim/coc-settings.json
+++ .vim/coc-settings.json
@@ ... @@
   // eslint settings
   // Requires coc-eslint extension to be installed
-  "eslint.enable": true,
+  "eslint.filetypes": ["javascript", "javascriptreact", "typescript", "typescriptreact"],
   "eslint.packageManager": "npm",
   "eslint.autoFixOnSave": true,
@@ ... @@
```

This change specifies that ESLint should only be enabled for `javascript`, `javascriptreact`, `typescript`, and `typescriptreact` file types. If there are other file types you want ESLint to handle, you can add them to the `eslint.filetypes` array.